% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRAMP.R
\name{BRAMP}
\alias{BRAMP}
\title{This is the main BRAMP function.}
\usage{
BRAMP(y = NULL, X = NULL, y_SAC_node = NULL, xlocs = 0, ylocs = 0,
  nr_iterations = 0, chain_thinout = 10, result_file = NULL,
  mcmc_rnd_seed = NULL, fixed_edges = NULL, edge_fanin = 5)
}
\arguments{
\item{y}{m-length vector with the number of 'm' observations for the target node.}

\item{X}{n-by-m matrix with 'n' nodes and 'm' observations.}

\item{y_SAC_node}{m-length vector with the spatial autocorrelation data. Can be left NULL if not used.}

\item{xlocs}{Integer defining the number of observations along the x-axis.}

\item{ylocs}{Integer defining the number of observations along the x-axis. Note that xlocs * ylocs == m.}

\item{nr_iterations}{Number of MCMC iterations to run.}

\item{chain_thinout}{Save samples from the chain every 'chain_thinout' iteration.}

\item{result_file}{A result file from a previous MCMC simulation that should be continued. If 'nr_iterations' is greater than the iterations in the result file, the simulation is continued until 'nr_iterations'.}

\item{mcmc_rnd_seed}{Integer value specifying a random seed.}

\item{fixed_edges}{Vector of edge indicators that should not be change by an edge move.}

\item{edge_fanin}{Maximum number of incoming edges for the response.}
}
\description{
This is the main BRAMP function.
}
