% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_main.R
\name{mcmc_main}
\alias{mcmc_main}
\title{The main function that runs the MCMC simulation.}
\usage{
mcmc_main(y, X, MCMC.chain, Grid.obj, HYPERvar, start.iter, end.iter)
}
\description{
The main function that runs the MCMC simulation.
}
