% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{readInput}
\alias{readInput}
\title{Read input data, either as matrix or as file name.}
\usage{
readInput(data)
}
\arguments{
\item{data}{Path name to file or matrix or data.frame to be read.}
}
\description{
Read input data, either as matrix or as file name.
}
