% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mondrian_moves.R
\name{shift.cut}
\alias{shift.cut}
\title{Shift an existing cut in a leaf pair.
This moves just repositions the cut, i.e. x or y and locations in the same 
manner we would set a new cut. 
FIXME: Maybe we could create two moves: one that slightly shifts a cut along 
his orientation another that changes the orientation: x -> y, y -> x        
In all cases, the prior should not change because halfperimeter and cost of 
cut stay the same.}
\usage{
shift.cut(Grid.obj, X, Y, HYPERvar, DEBUGLVL = 0, counter = 0)
}
\description{
Shift an existing cut in a leaf pair.
This moves just repositions the cut, i.e. x or y and locations in the same 
manner we would set a new cut. 
FIXME: Maybe we could create two moves: one that slightly shifts a cut along 
his orientation another that changes the orientation: x -> y, y -> x        
In all cases, the prior should not change because halfperimeter and cost of 
cut stay the same.
}
